requirements.txt
Flask>=2.0
requests>=2.25
Pillow>=8.0
flask-cors>=3.0
